Fix issue #1
Collision detection is now perfect, or a bit late. This is better than
detecting collisions too early.

I moved the isTouching check from the main program into the cactus. In
this way, each cactus is responsible for checking whether it is
colliding with the cactus. This saves us from having repeated checks
inside the main program.